refactor synctank and tankstats to service
